feat: add modular configuration management and UI templates for Ollama settings
